tg_bot/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Events==0.5
fastapi==0.115.4
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
opensearch-py==2.7.1
packaging==24.2
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
transformers==4.46.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0

python-dotenv
pi-heif
pikepdf
aiogram==3.6.0
celery==5.2.3
redis==4.5.5
unstructured==0.16.5
unstructured-inference==0.8.1
python-dotenv==0.20.0